requirements.txt
torch
torchvision
numpy
matplotlib
seaborn
scikit-learn
pillow
gradio
libmr